+++ app/src/Exceptions/ValidationException.php
@@ -11,6 +11,7 @@
 namespace UserFrosting\Sprinkle\Core\Exceptions;
 
 use Exception;
+use Throwable;
 use UserFrosting\Sprinkle\Core\Exceptions\Contracts\UserMessageException;
 use UserFrosting\Support\Message\UserMessage;
 
@@ -23,6 +24,14 @@ final class ValidationException extends Exception implements UserMessageExceptio
      * @var string[]
      */
     protected array $errors = [];
+
+    /**
+     * {@inheritDoc}
+     */
+    public function __construct(string $message = '', int $code = 400, ?Throwable $previous = null)
+    {
+        parent::__construct($message, $code, $previous);
+    }
 
     /**
      * Add errors returned by Valitron/Validator.