+++ app/src/ServicesProvider/ErrorHandlerService.php
@@ -18,10 +18,8 @@ use UserFrosting\ServicesProvider\ServicesProviderInterface;
 use UserFrosting\Sprinkle\Core\Error\ExceptionHandlerMiddleware;
 use UserFrosting\Sprinkle\Core\Error\Handler\HttpExceptionHandler;
 use UserFrosting\Sprinkle\Core\Error\Handler\PhpMailerExceptionHandler;
-use UserFrosting\Sprinkle\Core\Error\Handler\UserFacingExceptionHandler;
-use UserFrosting\Sprinkle\Core\Error\Handler\ValidationExceptionHandler;
-use UserFrosting\Sprinkle\Core\Exceptions\UserFacingException;
-use UserFrosting\Sprinkle\Core\Exceptions\ValidationException;
+use UserFrosting\Sprinkle\Core\Error\Handler\UserMessageExceptionHandler;
+use UserFrosting\Sprinkle\Core\Exceptions\Contracts\UserMessageException;
 
 class ErrorHandlerService implements ServicesProviderInterface
 {
@@ -33,9 +31,8 @@ class ErrorHandlerService implements ServicesProviderInterface
              */
             ExceptionHandlerMiddleware::class => \DI\autowire()
                 ->method('registerHandler', HttpException::class, HttpExceptionHandler::class, true)
-                ->method('registerHandler', UserFacingException::class, UserFacingExceptionHandler::class, true)
-                ->method('registerHandler', PHPMailerException::class, PhpMailerExceptionHandler::class)
-                ->method('registerHandler', ValidationException::class, ValidationExceptionHandler::class),
+                ->method('registerHandler', UserMessageException::class, UserMessageExceptionHandler::class, true)
+                ->method('registerHandler', PHPMailerException::class, PhpMailerExceptionHandler::class),
         ];
     }
 }