+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [5.0.0-alpha4](https://github.com/userfrosting/sprinkle-core/compare/5.0.0-alpha4...5.0.0-alpha5)
+
+- [Exceptions] `SprunjeException` now extends `UserFacingException`.
+- [Exceptions] `UserFacingExceptionHandler` renamed `UserMessageExceptionHandler`.
+- [Exceptions] `ValidationExceptionHandler` replaced removed (`ValidationException` now handled by `UserMessageExceptionHandler`).
+- [Exceptions] `UserMessageExceptionHandler` now add all `UserMessageException` interface instead of only `UserFacingException` class to alert stream and is used to handle all `UserMessageException` instead of `UserFacingException`.
+
 ## [5.0.0-alpha4](https://github.com/userfrosting/sprinkle-core/compare/5.0.0-alpha3...5.0.0-alpha4)
 
 - [Exceptions] Added base `UserFacingException` and `NotFoundException`.